# swagger.yaml
...
swagger: '2.0'
info:
  title: IoT Device Repository
  termsOfService: http://swagger.io/terms/
  description: 'The IOT Device Repository provides metadata for all IOT devices managed on the platform, their device classes and the associated data structures for the user.'
  version: '1.0'
  contact:
    name: sh20xyqi
    email: sh20xyqi@studserv.uni-leipzig.de
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
schemes:
  - https
host: api.sepl.infai.org
basePath: /iot-device-repo
produces:
  - text/plain
...
paths:
...
  /deviceType:
    post:
      tags:
        - deviceTypeApi
      summary:
        Returns a sorted list containing all device instances.
      description:
        ''
      produces:
        - application/json
      parameters:
      - in: body
        name: body
        description: Device type object that needs to be added to the database.
        required: true
        schema:
          $ref: '#/definitions/DeviceType'
      responses:
        200:
          description: Returns the instanciated device type.
          schema:
            $ref: '#/definitions/DeviceType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - JSON encoding failed; DeviceType is not valid;  DeviceType is not consistent; DeviceType can not be instanciated.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /deviceType/{id}:
    get:
      tags:
        - deviceTypeApi
      summary:
        Returns a sorted list containing all device instances.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the to be returned device type.
          required: true
          type: string
      responses:
        200:
          description: Returns the device type that belongs to the id.
          schema:
            $ref: '#/definitions/DeviceType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceType can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
    post:
      tags:
        - deviceTypeApi
      summary:
        Updates device type by Id.
      description:
        ''
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the to be modified device type.
          required: true
          type: string
        - name: body
          in: body
          description: Device type object that needs to be updated.
          required: true
          schema:
            $ref: '#/definitions/DeviceType'
      responses:
        200:
          description: Returns the modified device type.
          schema:
            $ref: '#/definitions/DeviceType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceType is not valid;  DeviceType is not consistent; DeviceType can not be instanciated.
          schema:
            $ref: '#/definitions/ErrorMessage'
    delete:
      tags:
        - deviceTypeApi
      summary:
        Deletes device type by Id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the the to be deleted device type.
          required: true
          type: string
      responses:
        200:
          description: Deletes the device type behind the passed ID.
          schema:
            $ref: '#/definitions/DeviceType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - DeviceType can not be found in database; dependend device instances exist; DeviceType can not be deleted.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /deviceType/{id}/{depth}:
    get:
      tags:
        - deviceTypeApi
      summary:
        ''
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: Identifies the the to be returned device type.
          required: true
          type: string
        - name: depth
          in: path
          description: ''
          required: true
          type: integer
      responses:
        200:
          description: ''
          schema:
            $ref: '#/definitions/DeviceType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - Depth can not be found in database or can not be converted into string; Id and depth results an error when getting device type from database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /deviceTypes/{limit}/{offset}:
    get:
      tags:
        - deviceTypeApi
      summary:
        Returns device types within a limit and how much elements had to be skipped.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: limit
          in: path
          description: Restricts the set of returnable elements to the passed limit.
          required: true
          type: integer
        - name: offset
          in: path
          description: Specifies how many elements of the set are skipped.
          required: true
          type: integer
      responses:
        200:
          description: Returns a sorted list of device types.
          schema:
            type: array
            items:
              $ref: '#/definitions/DeviceType'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - A specific limit can not be found; a specific offset can not be found; a DeviceDeviceList can not be returned.
          schema:
            $ref: '#/definitions/ErrorMessage'
  /service/{id}:
    get:
      tags:
        - deviceTypeApi
      summary:
        Returns service by Id.
      description:
        ''
      produces:
        - application/json
      parameters:
        - name: id
          in: path
          description: identifies the to be returned service.
          required: true
          type: string
      responses:
        200:
          description: Returns the service behind the passed Id.
          schema:
            $ref: '#/definitions/Service'
        500:
          description: Returns an ErrorMessage containing further information about the specific error. This error could be - Service can not be found in database.
          schema:
            $ref: '#/definitions/ErrorMessage'
  #Everything that belongs to deviceTypeApi/guiApi.go
...
definitions:
  MsgSegment:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      constraints:
        type: array
        items:
          type: string
    required:
      - id
      - name
      - constraints
  Protocol:
    type: object
    properties:
      id:
        type: string
        format: string
      protocol_handler_url:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      msg_structure:
        type: array
        items:
          $ref: '#/definitions/MsgSegment'
    required:
      - id
      - protocol_handler_url
      - name
      - description
      - msg_structure
...
  Service:
    type: object
    properties:
      id:
        type: string
        format: string
      service_type:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      protocol:
        type: object
        format: '#/definitions/Protocol'
      input:
        type: array
        items:
          $ref: '#/definitions/TypeAssignment'
      output:
        type: array
        items:
          $ref: '#/definitions/TypeAssignment'
      url:
        type: string
        format: string
    required:
      - id
      - service_type
      - name
      - description
      - protocol
      - input
      - output
      - url
  AdditionalFormatInfo:
    type: object
    properties:
      id:
        type: string
        format: string
      field:
        type: object
        format: '#/definitions/FieldType'
      format_flag:
        type: string
        format: string
    required:
      - id
      - field
      - format_flag
  TypeAssignment:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      msg_segment:
        type: object
        format: '#/definitions/MsgSegment'
      type:
        type: object
        format: '#/definitions/ValueType'
      format:
        type: string
        format: string
      additional_formatinfo:
        type: array
        items:
          $ref: '#/definitions/AdditionalFormatInfo'
    required:
      - id
      - name
      - msg_segment
      - type
      - format
      - additional_formatinfo
  DeviceType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      device_class:
        type: object
        format: '#/definitions/DeviceClass'
      services:
        type: array
        items:
          $ref: '#/definitions/Service'
      vendor:
        type: object
        format: '#/definitions/Vendor'
      config_parameter:
        type: array
        items:
          $ref: '#/definitions/ConfigFieldType'
    required:
      - id
      - name
      - description
      - device_class
      - services
      - vendor
      - config_parameter
...
  Vendor:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
    required:
      - id
      - name
  DeviceClass:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
    required:
      - id
      - name
...
  ConfigFieldType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
    required:
      - id
      - name
  FieldType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      type:
        type: object
        format: '#/definitions/ValueType'
    required:
      - id
      - name
      - type
  ValueType:
    type: object
    properties:
      id:
        type: string
        format: string
      name:
        type: string
        format: string
      description:
        type: string
        format: string
      base_type:
        type: string
        format: string
      fields:
        type: array
        items:
          $ref: '#/definitions/FieldType'
      literal:
        type: string
        format: string
    required:
      - id
      - name
      - description
      - base_type
      - fields
      - literal
...
  ErrorMessage:
    type: object
    properties:
      status_code:
        type: integer
        format: int32
      message:
        type: string
        format: string
      error_code:
        type: string
        format: string
      detail:
        type: array
        items:
          type: string
    required:
      - status_code
      - message
      - error_code
      - detail
...